Stream: fixed possible integer overflow in rate limiting.

+++ src/stream/ngx_stream_proxy_module.c
@@ -1062,7 +1062,7 @@ ngx_stream_proxy_process(ngx_stream_session_t *s, ngx_uint_t from_upstream,
                     break;
                 }
 
-                if (size > (size_t) limit) {
+                if ((off_t) size > limit) {
                     size = (size_t) limit;
                 }
             }
